+++ R/lma_simets.R
@@ -185,7 +185,13 @@ lma_simets <- function(a, b = NULL, metric = NULL, group = NULL, lag = 0, agg = 
     if ((pairwise && symmetrical) || mean) {
       for (i in seq_along(res)) {
         if (pairwise && (symmetrical || mean)) res[[i]] <- forceSymmetric(res[[i]], "L")
-        if (mean) res[[i]] <- if (is.null(dim(res[[i]]))) mean(res[[i]], na.rm = TRUE) else rowMeans(res[[i]], TRUE)
+        if (mean) {
+          res[[i]] <- if (is.null(dim(res[[i]]))) {
+            mean(res[[i]], na.rm = TRUE)
+          } else {
+            (rowSums(res[[i]], TRUE) - 1) / (ncol(res[[i]]) - 1)
+          }
+        }
       }
     }
     if (is.null(dim(res[[1]]))) {